refactor: split gc/layout.rs and codegen/artifacts.rs under the 2000-line cap
#8204 pushed both files over scripts/check_file_size.sh's hard cap
(layout.rs 1975 -> 2110, artifacts.rs 2000 -> 2005), turning the required
lint context red on main for every PR. Pure code moves, no logic change:

- gc/layout.rs: the typed-shape layout installation protocol
  (TypedShapeProof, mask_words, init_typed_shape_layout,
  install_typed_shape_layout_slow, typed_shape_layout_entry,
  js_gc_init_typed_shape_layout, js_gc_declare_typed_shape_layout) moves
  to gc/layout/typed_shape.rs, next to the existing layout/slot_mask.rs.
  2110 -> 1778 lines. The two extern "C" entry points keep their
  crate::gc:: paths via an explicit named re-export.

- codegen/artifacts.rs: synthesized_ctor_param_count moves to a new
  sibling codegen/ctor_arity.rs. 2005 -> 1930 lines.
